requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file .\requirements.txt .\requirements.in
#
django-simple-elasticsearch==2.0.0
django==1.10.5
elasticsearch==2.4.1      # via django-simple-elasticsearch
urllib3==1.20             # via elasticsearch
